Added version info to the Manage dialog.
