+++ Source/ExcelDna.AddInManager/Dialogs/AddinsListView.xaml
@@ -1,0 +1,18 @@
+﻿<UserControl x:Class="ExcelDna.AddInManager.AddinsListView" x:ClassModifier="internal"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:ExcelDna.AddInManager"
+             mc:Ignorable="d" 
+             d:DesignHeight="450" d:DesignWidth="800">
+    <ListView x:Name="addinsListView">
+        <ListView.View>
+            <GridView>
+                <GridViewColumn Header="Name" DisplayMemberBinding="{Binding Path=ProductName}" />
+                <GridViewColumn Header="Company" DisplayMemberBinding="{Binding Path=CompanyName}" />
+                <GridViewColumn Header="Version" DisplayMemberBinding="{Binding Path=Version}" />
+            </GridView>
+        </ListView.View>
+    </ListView>
+</UserControl>
