+++ Source/ExcelDna.AddInManager/Dialogs/InstallDialog.xaml.cs
@@ -1,31 +1,14 @@
-﻿using System.IO;
-using System.Windows;
+﻿using System.Windows;
 
 namespace ExcelDna.AddInManager
 {
     internal partial class InstallDialog : Window
     {
-        private class ListItem
-        {
-            public ListItem(AddInVersionInfo addin)
-            {
-                this.Addin = addin;
-            }
-
-            public string? CompanyName => Addin.CompanyName;
-            public string? ProductName => Addin.IsVersioned ? Addin.ProductName : Path.GetFileNameWithoutExtension(Addin.Path);
-            public string? Version => Addin.Version?.ToString();
-
-            public AddInVersionInfo Addin { get; }
-        }
-
         public InstallDialog(List<AddInVersionInfo> addins)
         {
             InitializeComponent();
-            this.addins = addins;
 
-            foreach (var i in addins)
-                addinsListView.Items.Add(new ListItem(i));
+            addinsListView.Add(addins);
         }
 
         public List<AddInVersionInfo>? GetSelectedAddins()
@@ -37,10 +20,7 @@ namespace ExcelDna.AddInManager
         {
             try
             {
-                selectedAddins = new List<AddInVersionInfo>();
-                int i = addinsListView.SelectedIndex;
-                if (i >= 0)
-                    selectedAddins.Add(addins[i]);
+                selectedAddins = addinsListView.GetSelectedAddins();
 
                 DialogResult = true;
             }
@@ -50,7 +30,6 @@ namespace ExcelDna.AddInManager
             }
         }
 
-        private List<AddInVersionInfo> addins;
         private List<AddInVersionInfo>? selectedAddins;
     }
 }
