+++ Source/ExcelDna.AddInManager/Dialogs/ManageDialog.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="ExcelDna.AddInManager.ManageDialog"
+﻿<Window x:Class="ExcelDna.AddInManager.ManageDialog" x:ClassModifier="internal"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
@@ -11,7 +11,7 @@
             <ColumnDefinition Width="*"/>
             <ColumnDefinition Width="Auto"/>
         </Grid.ColumnDefinitions>
-        <ListBox Grid.Column="0" Margin="0,0,11,0" x:Name="addinsListBox"/>
+        <local:AddinsListView Grid.Column="0" Margin="0,0,11,0" x:Name="addinsListView"/>
         <Button Grid.Column="1" Name="uninstallButton" VerticalAlignment="Top" MinWidth="70" Click="OnUninstall">Uninstall</Button>
     </Grid>
 </Window>
